perf(usenet): in-process VFS reads, zero-copy delivery, leaner allocs
Collapse the usenet serving path and shave the remaining per-read cost.

Architecture:
- Remove the loopback `/usenet/{info_hash}/{file_index}` HTTP route. The
  VFS now reads usenet content in-process via a `LocalByteSource` trait,
  eliminating the FUSE -> HTTP -> streamer hop and its duplicate buffering.
- Identify usenet entries by explicit `usenet_info_hash`/`usenet_file_index`
  columns (migration 029, with a regex backfill of legacy loopback URLs)
  instead of parsing a fake stream URL. The plugin now emits a self-
  contained `usenet://{hash}/{index}` marker.
- Drop the now-unused public base URL setting; move the active-streams
  "now playing" registry into the VFS session.
- Remove the legacy `…/usenet/…` URL parsing/fallback entirely now that
  every row carries the explicit columns.

Streaming hot path:
- UsenetFrameReader consumes the streamer's decoded `Bytes` frames directly
  (zero-copy `split_to` for sub-frame reads, single exactly-sized concat for
  straddles), replacing the BufReader/StreamReader/Vec double-copy sandwich.
- Read into uninitialized spare capacity (`read_buf`) in the NNTP fill loop
  and the HTTP chunk reader, dropping per-read/-article memsets.
- Key the segment cache, in-flight map, and decoded-size cache on `Arc<str>`
  so a cold fetch allocates the message-id once instead of ~4 times; add a
  non-cloning `SegmentCache::contains` for prefetch filtering.

Also: zero remaining clippy warnings across the workspace (restriction-lint
cleanups: clamp, uninit reads, named must-use ignores, graceful errors in
place of unreachable!).

Deployed and verified: migration backfills 1082/1082 entries, sequential +
seek reads byte-identical across cold/cache-hit, sustained reads stall-free,
memory bounded ~190-350 MB (was ~2 GB), no panics/EIO/FUSE errors.

+++ crates/riven-usenet/src/rar/v5.rs
@@ -279,7 +279,7 @@ pub(super) fn block_layout_v5(bytes: &[u8]) -> Option<(u64, u64)> {
     }
     let mut pos = 4; // skip CRC32
     let size_vint_start = pos;
-    let header_size = read_vint(bytes, &mut pos)? as u64;
+    let header_size = read_vint(bytes, &mut pos)?;
     let header_total = pos as u64 + header_size;
 
     // Walk far enough into the header to determine the data-area size.
